tests: move End2End tests in End2End\Tests\ namespace
