+++ server/tests/End2End/Administration/Infrastructure/Settings/Controller/PostSettingsControllerTest.php
@@ -11,19 +11,48 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Unit\Tests\Administration\Infrastructure\Settings\Controller;
+namespace End2End\Tests\Administration\Infrastructure\Settings\Controller;
 
+use End2End\Tests\AbstractControllerTest;
+use End2End\Tests\DatabaseHelper;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Unit\Tests\AbstractControllerTest;
-use Unit\Tests\DatabaseHelper;
 
-class PutSettingsControllerTest extends AbstractControllerTest
+class PostSettingsControllerTest extends AbstractControllerTest
 {
     /**
      * @throws \JsonException
      */
-    final public function testPutSettingsFailWithBadUuid(): void
+    final public function testPostSettingsSuccess(): void
+    {
+        // Arrange
+        DatabaseHelper::loadFixtures([['group' => 'user']]);
+        $content = [
+            'currency' => 'Euro',
+            'locale' => 'Fr',
+        ];
+        $adminClient = $this->createAdminClient();
+        $adminClient->request(
+            Request::METHOD_POST,
+            '/api/administration/settings/',
+            [],
+            [],
+            ['CONTENT_TYPE' => 'application/json'],
+            \json_encode($content, \JSON_THROW_ON_ERROR)
+        );
+
+        // Act
+        $response = $this->client->getResponse();
+
+        // Assert
+        self::assertSame(Response::HTTP_CREATED, $response->getStatusCode());
+        self::assertSame('Settings created started!', $response->getContent());
+    }
+
+    /**
+     * @throws \JsonException
+     */
+    final public function testPostSettingsAlreadyExist(): void
     {
         // Arrange
         DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'settings']]);
@@ -35,8 +64,8 @@ class PutSettingsControllerTest extends AbstractControllerTest
         // Act
         $adminClient = $this->createAdminClient();
         $adminClient->request(
-            Request::METHOD_PUT,
-            '/api/administration/settings/626adfca-fc5d-415c-9b7a-7541030bd147',
+            Request::METHOD_POST,
+            '/api/administration/settings/',
             [],
             [],
             ['CONTENT_TYPE' => 'application/json'],
@@ -46,34 +75,5 @@ class PutSettingsControllerTest extends AbstractControllerTest
 
         // Assert
         self::assertSame(Response::HTTP_INTERNAL_SERVER_ERROR, $response->getStatusCode());
-    }
-
-    /**
-     * @throws \JsonException
-     */
-    final public function testPutSettingsSuccess(): void
-    {
-        // Arrange
-        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'settings']]);
-        $content = [
-            'currency' => 'Euro',
-            'locale' => 'Fr',
-        ];
-
-        // Act
-        $adminClient = $this->createAdminClient();
-        $adminClient->request(
-            Request::METHOD_PUT,
-            '/api/administration/settings/a136c6fe-8f6e-45ed-91bc-586374791033',
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            \json_encode($content, \JSON_THROW_ON_ERROR)
-        );
-        $response = $this->client->getResponse();
-
-        // Assert
-        self::assertSame(Response::HTTP_ACCEPTED, $response->getStatusCode());
-        self::assertSame('Settings updated started!', $response->getContent());
     }
 }
