+++ server/tests/End2End/Administration/Infrastructure/Supplier/Controller/DeleteSupplierControllerTest.php
@@ -11,25 +11,26 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Unit\Tests\Administration\Infrastructure\Company\Controller;
+namespace End2End\Tests\Administration\Infrastructure\Supplier\Controller;
 
+use End2End\Tests\AbstractControllerTest;
+use End2End\Tests\DatabaseHelper;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Unit\Tests\AbstractControllerTest;
-use Unit\Tests\DatabaseHelper;
 
-class DeleteCompanyControllerTest extends AbstractControllerTest
+class DeleteSupplierControllerTest extends AbstractControllerTest
 {
     /**
      * @throws \JsonException
      */
-    final public function testDeleteCompanyFailWithBadUuid(): void
+    final public function testDeleteSupplierFailWithBadUuid(): void
     {
         // Arrange
-        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'company']]);
+        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'supplier']]);
         $adminClient = $this->createAdminClient();
 
         // Act
-        $adminClient->request('DELETE', '/api/administration/companies/626adfca-fc5d-415c-9b7a-7541030bd147');
+        $adminClient->request('DELETE', '/api/administration/suppliers/626adfca-fc5d-415c-9b7a-7541030bd147');
         $response = $adminClient->getResponse();
 
         // Assert
@@ -39,15 +40,18 @@ class DeleteCompanyControllerTest extends AbstractControllerTest
     /**
      * @throws \JsonException
      */
-    final public function testDeleteCompanySuccess(): void
+    final public function testDeleteSupplierSuccess(): void
     {
         // Arrange
-        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'company']]);
+        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'supplier']]);
         $adminClient = $this->createAdminClient();
-        $adminClient->request('DELETE', '/api/administration/companies/a136c6fe-8f6e-45ed-91bc-586374791033');
+        $adminClient->request(
+            Request::METHOD_DELETE,
+            '/api/administration/suppliers/a136c6fe-8f6e-45ed-91bc-586374791033'
+        );
 
         // Act
-        $response = $this->client->getResponse();
+        $response = $adminClient->getResponse();
 
         // Assert
         self::assertSame(Response::HTTP_NO_CONTENT, $response->getStatusCode());
