+++ server/tests/End2End/Administration/Infrastructure/Supplier/Controller/GetSuppliersControllerTest.php
@@ -11,24 +11,24 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Unit\Tests\Administration\Infrastructure\Settings\Controller;
+namespace End2End\Tests\Administration\Infrastructure\Supplier\Controller;
 
+use End2End\Tests\AbstractControllerTest;
+use End2End\Tests\DatabaseHelper;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Unit\Tests\AbstractControllerTest;
-use Unit\Tests\DatabaseHelper;
 
-class GetSettingsControllerTest extends AbstractControllerTest
+class GetSuppliersControllerTest extends AbstractControllerTest
 {
     /**
      * @throws \JsonException
      */
-    final public function testGetSettingsNoData(): void
+    final public function testGetSupplierNoData(): void
     {
         // Arrange
         DatabaseHelper::loadFixtures([['group' => 'user']]);
         $adminClient = $this->createAdminClient();
-        $adminClient->request(Request::METHOD_GET, '/api/administration/settings/');
+        $adminClient->request(Request::METHOD_GET, '/api/administration/suppliers/');
 
         // Act
         $response = $adminClient->getResponse();
@@ -38,12 +38,15 @@ class GetSettingsControllerTest extends AbstractControllerTest
         self::assertSame('No data found', $response->getContent());
     }
 
-    final public function testGetSettingsSuccess(): void
+    /**
+     * @throws \JsonException
+     */
+    final public function testGetSuppliersSuccess(): void
     {
         // Arrange
-        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'settings']]);
+        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'supplier']]);
         $adminClient = $this->createAdminClient();
-        $adminClient->request(Request::METHOD_GET, '/api/administration/settings/');
+        $adminClient->request(Request::METHOD_GET, '/api/administration/suppliers/');
 
         // Act
         $response = $adminClient->getResponse();
@@ -52,6 +55,6 @@ class GetSettingsControllerTest extends AbstractControllerTest
         self::assertSame(Response::HTTP_OK, $response->getStatusCode());
         $content = \json_decode($response->getContent(), true, 512, \JSON_THROW_ON_ERROR);
         self::assertIsArray($content);
-        self::assertEquals('a136c6fe-8f6e-45ed-91bc-586374791033', $content['uuid']);
+        self::assertEquals('a136c6fe-8f6e-45ed-91bc-586374791033', $content[0]['uuid']);
     }
 }
