+++ server/tests/End2End/Administration/Infrastructure/User/Controller/GetUsersControllerTest.php
@@ -11,42 +11,24 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Unit\Tests\Administration\Infrastructure\Company\Controller;
+namespace End2End\Tests\Administration\Infrastructure\User\Controller;
 
+use End2End\Tests\AbstractControllerTest;
+use End2End\Tests\DatabaseHelper;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Unit\Tests\AbstractControllerTest;
-use Unit\Tests\DatabaseHelper;
 
-class GetCompaniesControllerTest extends AbstractControllerTest
+class GetUsersControllerTest extends AbstractControllerTest
 {
     /**
      * @throws \JsonException
      */
-    final public function testGetCompaniesNoData(): void
+    final public function testGetUsersSuccess(): void
     {
         // Arrange
         DatabaseHelper::loadFixtures([['group' => 'user']]);
         $adminClient = $this->createAdminClient();
-        $adminClient->request(Request::METHOD_GET, '/api/administration/companies/');
-
-        // Act
-        $response = $adminClient->getResponse();
-
-        // Assert
-        self::assertSame(Response::HTTP_ACCEPTED, $response->getStatusCode());
-        self::assertSame('No data found', $response->getContent());
-    }
-
-    /**
-     * @throws \JsonException
-     */
-    final public function testGetCompaniesSuccess(): void
-    {
-        // Arrange
-        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'company']]);
-        $adminClient = $this->createAdminClient();
-        $adminClient->request(Request::METHOD_GET, '/api/administration/companies/');
+        $adminClient->request(Request::METHOD_GET, '/api/administration/users/');
 
         // Act
         $response = $adminClient->getResponse();
