+++ server/tests/End2End/DatabaseHelper.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Unit\Tests;
+namespace End2End\Tests;
 
 use Symfony\Component\Process\Process;
